identity-delete-trigger-py/requirements.txt
functions-framework==3.*
google-cloud-firestore==2.11.1
google-cloud-logging